test(host): establish gtest structure and initial nvs_core unit tests
- Setup GoogleTest/GoogleMock infrastructure in host_test/
- Fix lcov coverage patterns to support absolute paths
- Implement initial NvsCore tests:
    * Partition initialization with erase/retry logic
    * Data loading and committing
    * Schema version migration/validation
    * Error propagation on open and load
    * Full factory reset sequence

+++ app_water_tank/main/src/water_tank_nvs.cpp
@@ -25,6 +25,6 @@ void WaterTankNvs::setAppDefaults()
     stats.reset();
     
     // Core identity defaults
-    core_.node_id = static_cast<uint8_t>(FarmNodeId::WATER_TANK);
-    core_.node_type = static_cast<uint8_t>(FarmNodeType::SENSOR);
+    core_.node_id = FarmNodeId::WATER_TANK;
+    core_.node_type = FarmNodeType::SENSOR;
 }
